backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.2
Django==4.0.3
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
mutagen==1.45.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
zipp==3.8.0
